+++ Math/Combination/Pascal.cpp
@@ -1,9 +1,9 @@
-uint_fast64_t comb(uint_fast64_t n, uint_fast32_t r) {
+unsigned long long comb(unsigned long long n, unsigned int r) {
     if (r == 0) return 1;
     if (n < 0) return comb(-n + r - 1, r) * (r % 2 ? -1 : 1);
     if (n == 0 || r < 0 || n < r) return 0;
-    uint_fast64_t ret = 1;
-    for (uint_fast64_t i = 1; i <= r; ++i) (ret *= n--) /= i;
+    unsigned long long ret = 1;
+    for (unsigned long long i = 1; i <= r; ++i) (ret *= n--) /= i;
     return ret;
 }
