+++ Math/Prime/PrimeCheck.cpp
@@ -1,4 +1,4 @@
-bool isPrime(int_fast64_t n) {
+bool isPrime(long long n) {
     if (n < 2) return false;
     for (int i = 2; i * i <= n; ++i) if (n % i == 0) return false;
     return true;
